% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{call_snps}
\alias{call_snps}
\title{Call Single Nucleotide Polymorphisms (SNPs)}
\usage{
call_snps(dnam, non_carrier_threshold = 0.2,
  homozygous_threshold = 0.8, plot = TRUE)
}
\arguments{
\item{dnam}{list-output of \link{preprocess}}

\item{non_carrier_threshold}{SNPs with theta below this will be classified as non-carrier (default 0.2)}

\item{homozygous_threshold}{SNPs with theta above this will be classified as homozygous (default 0.8)}

\item{plot}{\code{TRUE} shows box plot-overview of theta-values among the three classes for all SNPs}
}
\value{
Returns matrix with samples in rows and SNPs in columns; 0 for non-carrier, 1 for heterozygous, 2 for homozygous
}
\description{
Even though the Illumina Infinium Bead Chips are made for CpG sites, they also contain a few (~60) highly polymorphic SNPs to help identify replicates. This function "calls" these SNPs, meaning it defines them as either non-carrier (0), heterozygous (1) or homozygous (2) based on the theta-value in the polar coordinate system.
}
\seealso{
\link{preprocess}, \link{snp_distance}
}
