% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{estimate_leukocytes}
\alias{estimate_leukocytes}
\title{Estimate leukocyte proportions}
\usage{
estimate_leukocytes(dnam)
}
\arguments{
\item{dnam}{list-output of \link{preprocess}}
}
\value{
Returns a dataframe with samples in rows and leukocyte-subtype-proportions in columns.
}
\description{
Estimates white-blood-cell proportions from blood samples using the Houseman method and coefficients derived from the Reinius dataset. The following subtypes are estimated: Neutrophils, eosinophils, mast cells, CD4 T-cells, CD8 T-cells, B-cells, NK-cells
}
\references{
Houseman, Eugene Andres, et al. "DNA Methylation Arrays as Surrogate Measures of Cell Mixture Distribution." BMC Bioinformatics, vol. 13, no. 1, 2012, doi:10.1186/1471-2105-13-86.

Reinius, Lovisa E., et al. "Differential DNA Methylation in Purified Human Blood Cells: Implications for Cell Lineage and Studies on Disease Susceptibility." PLoS ONE, vol. 7, no. 7, 2012, doi:10.1371/journal.pone.0041361.
}
\seealso{
\link{preprocess}
}
