% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{identify_replicates}
\alias{identify_replicates}
\title{Identifies replicates}
\usage{
identify_replicates(dnam, snp_distance_margin = 0,
  kmeans = kmeans_default, centroids = centroids_default)
}
\arguments{
\item{dnam}{list-output of \link{preprocess}}

\item{snp_distance_margin}{specifies the maximum number of different alleles for two samples to still be considered replicate (default 0)}

\item{kmeans}{passed on to \link{infer_sex}}

\item{centroids}{passed on to \link{infer_sex}}
}
\value{
Returns dataframe with replicates
}
\description{
Two samples are considered replicates when their inferred sex is the same and their SNP-L1-distance doesn't exceed \code{snp_distance_margin}.
}
\seealso{
\link{preprocess}, \link{call_snps}, \link{snp_distance}, \link{infer_sex}
}
