% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{infer_sex}
\alias{infer_sex}
\title{Infer sex}
\usage{
infer_sex(dnam, kmeans = kmeans_default, centroids = centroids_default,
  plot = TRUE)
}
\arguments{
\item{dnam}{list-output of \link{preprocess}}

\item{kmeans}{\code{TRUE} performs k-means-clustering with predefined initialization. \code{FALSE} uses hard boundaries (\code{colMeans(centroids)}) which also yields NA values when uncertain.}

\item{centroids}{default initialization for k-means-clustering and boundaries}

\item{plot}{\code{TRUE} shows scatter plot with sex-inference}
}
\value{
Returns named factor-vector with levels "m" / "f" / NA (for \code{kmeans=FALSE} only)
}
\description{
Sex inference is performed based on two summary statistics: and median.chrX (higher for females because of X-chromosome-inactivation) and missing.chrY (higher for females but not 100% because of cross-hybridization of X-chromosomal-genes).
}
\seealso{
\link{preprocess}, \link{compare_sex}
}
