% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{linear_mixed_models}
\alias{linear_mixed_models}
\title{Performs multiple linear mixed models}
\usage{
linear_mixed_models(dnam, outcome, chip, position, confounders = NA,
  lrt = TRUE, save.residuals = FALSE, save.ranks = TRUE,
  plot = FALSE)
}
\arguments{
\item{dnam}{list-output of \link{preprocess} or generic methylation-beta-value-matrix}

\item{outcome}{named vector of outcome variable}

\item{chip}{only has to be provided when \code{dnam} is not a list-output of \link{preprocess}. Vector of chip id each sample was performed on.}

\item{position}{only has to be provided when \code{dnam} is not a list-output of \link{preprocess}. Vector of chip position each sample was performed on.}

\item{confounders}{optional dataframe with confounders, rownames matching those of dnam-list-entries. Even when only one confounder is present provide it as a 1-column-dataframe.}

\item{lrt}{whether p-values should be computed using likelihood-ratio tests. See ?omics::mlmer}

\item{save.residuals}{whether model residuals should be stored and returned. See ?omics::mlmer}

\item{save.ranks}{whether random effect ranks should be stored and returned. See ?omics::mlmer}

\item{plot}{\code{TRUE} plots two heatmap for the random effect ranks of chip id and position. See ?omics::ranks.heatmap}
}
\value{
A list with elements \code{coefficients} (containing \code{coef}, \code{coef.se} and \code{pval} for each CpG site) and optionally \code{residuals} and \code{ranef.ranks}. See ?omics::mlmer
}
\description{
One model is fitted per CpG site. Beta-values are treated as dependent variables and outcomes as independent variables. Chip id and position are treated as random effect intercepts in the mixed model to reduce technical variance. Additional confounders like age and sex can optionally be added and treated as fixed effects.
}
\seealso{
\link{preprocess}
}
