% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{snp_distance}
\alias{snp_distance}
\title{Computes genotype distance}
\usage{
snp_distance(dnam, plot_histogram = TRUE, plot_heatmap = FALSE)
}
\arguments{
\item{dnam}{list-output of \link{preprocess}}

\item{plot_histogram}{\code{TRUE} shows distribution of genotype-distance}

\item{plot_heatmap}{\code{TRUE} shows heatmap with all samples}
}
\value{
Returns \code{dist}-object which is lower triangle of distance matrix. To convert to matrix call \code{as.matrix()}
}
\description{
L1-distance-matrix of samples according to the ~60 SNPs present on Illumina Infinium Bead Chips corresponds to absolute number of different alleles.
}
\seealso{
\link{preprocess}, \link{call_snps}
}
