% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{remove_unreliable_samples_probes}
\alias{remove_unreliable_samples_probes}
\title{Removes unreliable samples and CpG sites}
\usage{
remove_unreliable_samples_probes(dnam, threshold_NA_sample = 0.1,
  threshold_NA_cpg = 0.2, bc_threshold = 1, plot = T)
}
\arguments{
\item{dnam}{list-output of \link{preprocess}}

\item{threshold_NA_sample}{samples (= rows) with more NAs than this threshold-fraction in their CpG matrix will be removed (default 0.1)}

\item{threshold_NA_cpg}{CpG sites (= columns) with more NAs than this threshold-fraction will be removed (default 0.2)}

\item{bc_threshold}{samples that failed bisulfite conversion (bc) according to this lower-threshold for bc-summary-statistics in samples dataframe will be removed (default 1)}

\item{plot}{\code{TRUE} for quality control plots showing NA fractions per row/column, bc-failure and average methylation beta-value-distribution among all samples}
}
\value{
Returns list in same format as \link{preprocess}
}
\description{
Samples are deemed unreliable if they have too many NA values in CpG matrix or have failed bisulfite conversion according to 3 summary statistics in samples dataframe (bc1.grn, bc1.red, bc2). CpG sites are deemed unreliable if they have too many NA values.
}
\references{
Illumina BeadArray Controls Reporter Software Guide (Document #1000000004009 v00, October 2015): https://support.illumina.com/content/dam/illumina-support/documents/documentation/chemistry_documentation/infinium_assays/infinium_hd_methylation/beadarray-controls-reporter-user-guide-1000000004009-00.pdf
}
\seealso{
\link{preprocess}
}
